+++ src/mplfe/common/include/feir_var.h
@@ -144,7 +144,7 @@ class FEIRVar {
     return EqualsToImpl(var);
   }
 
-  size_t Hash() const {
+  uint32 Hash() const {
     return HashImpl();
   }
 
@@ -177,7 +177,7 @@ class FEIRVar {
   virtual std::string GetNameRawImpl() const = 0;
   virtual std::unique_ptr<FEIRVar> CloneImpl() const = 0;
   virtual bool EqualsToImpl(const std::unique_ptr<FEIRVar> &var) const = 0;
-  virtual size_t HashImpl() const = 0;
+  virtual uint32 HashImpl() const = 0;
 
   FEIRVarKind kind : 6;
   bool isGlobal : 1;
